+++ allennlp/modules/seq2seq_encoders/__init__.py
@@ -6,17 +6,17 @@ others are AllenNLP modules.
 
 The available Seq2Seq encoders are
 
-- `"gru"` : https://pytorch.org/docs/master/nn.html#torch.nn.GRU
-- `"lstm"` : https://pytorch.org/docs/master/nn.html#torch.nn.LSTM
-- `"rnn"` : https://pytorch.org/docs/master/nn.html#torch.nn.RNN
-- `"augmented_lstm"` : allennlp.modules.augmented_lstm.AugmentedLstm
-- `"alternating_lstm"` : allennlp.modules.stacked_alternating_lstm.StackedAlternatingLstm
-- `"alternating_highway_lstm"` : allennlp.modules.stacked_alternating_lstm.StackedAlternatingLstm (GPU only)
-- `"stacked_self_attention"` : allennlp.modules.stacked_self_attention.StackedSelfAttentionEncoder
-- `"multi_head_self_attention"` : allennlp.modules.multi_head_self_attention.MultiHeadSelfAttention
-- `"pass_through"` : allennlp.modules.pass_through_encoder.PassThroughEncoder
-- `"feedforward"` : allennlp.modules.feedforward_encoder.FeedforwardEncoder
+- `"gru"` : allennlp.modules.seq2seq_encoders.GruSeq2SeqEncoder
+- `"lstm"` : allennlp.modules.seq2seq_encoders.LstmSeq2SeqEncoder
+- `"rnn"` : allennlp.modules.seq2seq_encoders.RnnSeq2SeqEncoder
+- `"augmented_lstm"` : allennlp.modules.seq2seq_encoders.AugmentedLstmSeq2SeqEncoder
+- `"alternating_lstm"` : allennlp.modules.seq2seq_encoders.StackedAlternatingLstmSeq2SeqEncoder
+- `"pass_through"` : allennlp.modules.seq2seq_encoders.PassThroughEncoder
+- `"feedforward"` : allennlp.modules.seq2seq_encoders.FeedForwardEncoder
 - `"pytorch_transformer"` : allennlp.modules.seq2seq_encoders.PytorchTransformer
+- `"compose"` : allennlp.modules.seq2seq_encoders.ComposeEncoder
+- `"gated-cnn-encoder"` : allennlp.momdules.seq2seq_encoders.GatedCnnEncoder
+- `"stacked_bidirectional_lstm"`: allennlp.modules.seq2seq_encoders.StackedBidirectionalLstmSeq2SeqEncoder
 """
 
 from allennlp.modules.seq2seq_encoders.compose_encoder import ComposeEncoder